refactor(dashboard): Extract shared locale-profiles module and fix spot-check batch counter
